+++ css/datatypes/functions.v
@@ -4,6 +4,6 @@ pub struct CalcSum {
 }
 
 // evaluate returns evaluates the sum and returns the value in pixels
-pub fn (sum CalcSum) evaluate(em f32, rem f32, vh f32, vw f32) f32 {
+pub fn (sum CalcSum) evaluate(em f64, rem f64, vh f64, vw f64) f64 {
 	return 0.0
 }
